requirements.txt
Django>=4.0.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1
arabic-reshaper==2.1.3
asgiref==3.4.1
beautifulsoup4==4.10.0
django-bootstrap-datepicker-plus>=3.0.5
django-bootstrap4>=21.1
django-crispy-forms>=1.13.0
django-widget-tweaks>=1.4.9
future==0.18.2
html5lib==1.1
Pillow==8.3.2
pycodestyle==2.7.0
PyPDF2==1.26.0
reportlab==3.6.5
python-bidi==0.4.2
pytz==2021.3
six==1.16.0
toml==0.10.2
sqlparse==0.4.2
xhtml2pdf==0.2.5
webencodings==0.5.1
soupsieve==2.3.1